+++ bindings/core/src/method_handler/utils.rs
@@ -64,9 +64,9 @@ pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response
         UtilsMethod::ComputeTokenId {
             alias_id,
             serial_number,
-            token_scheme_type,
+            token_scheme_kind,
         } => {
-            let foundry_id = FoundryId::build(&AliasAddress::new(alias_id), serial_number, token_scheme_type.kind());
+            let foundry_id = FoundryId::build(&AliasAddress::new(alias_id), serial_number, token_scheme_kind);
             Response::TokenId(TokenId::from(foundry_id))
         }
         UtilsMethod::HashTransactionEssence { essence } => {