+++ sdk/src/types/block/output/feature/metadata.rs
@@ -13,7 +13,6 @@ pub(crate) type MetadataFeatureLength =
 
 /// Defines metadata, arbitrary binary data, that will be stored in the output.
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, packable::Packable)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[packable(unpack_error = Error, with = |err| Error::InvalidMetadataFeatureLength(err.into_prefix_err().into()))]
 pub struct MetadataFeature(
     // Binary data.