+++ sdk/src/types/block/output/foundry.rs
@@ -280,7 +280,6 @@ impl From<&FoundryOutput> for FoundryOutputBuilder {
 
 /// Describes a foundry output that is controlled by an alias.
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct FoundryOutput {
     // Amount of IOTA tokens held by the output.
     amount: u64,